Exception Handling - delete

+++ src/main/java/com/edcandoweb/course/services/UserService.java
@@ -4,10 +4,12 @@ import java.util.List;
 import java.util.Optional;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.stereotype.Service;
 
 import com.edcandoweb.course.entities.User;
 import com.edcandoweb.course.repositories.UserRepository;
+import com.edcandoweb.course.services.exceptions.DataBaseException;
 import com.edcandoweb.course.services.exceptions.ResourceNotFoundException;
 
 @Service //Permite que seja injetado pelo autowired
@@ -30,8 +32,17 @@ public class UserService {
 	}
 	
 	public void delete(Long id) {
-		 userRepository.deleteById(id);
+	    
+		if (!userRepository.existsById(id)) {
+	        throw new ResourceNotFoundException(id);
+	    }
+	    try {
+		userRepository.deleteById(id);
+		}catch(DataIntegrityViolationException e) {
+			throw new DataBaseException(e.getMessage());
+		}
 	}
+
 	
 	public User update(Long id, User obj) {
 		User entity = userRepository.getReferenceById(id);
